BUG Fix: Chest collectors repairing items with high damage * Block bags were casting damage to (byte) causing value wrap * Updated all block bags to int for damage value

+++ src/AdminBlockSource.java
@@ -56,9 +56,9 @@ public class AdminBlockSource extends BlockBag {
      * @throws OutOfBlocksException
      */
     public void fetchBlock(int id) throws BlockSourceException {
-    	fetchBlock(id, (byte)-1);
+    	fetchBlock(id, -1);
     }
-    public void fetchBlock(int id, byte data) throws BlockSourceException {
+    public void fetchBlock(int id, int data) throws BlockSourceException {
         if (!canFetch) {
             throw new OutOfBlocksException(id);
         }
@@ -73,9 +73,9 @@ public class AdminBlockSource extends BlockBag {
      * @throws OutOfSpaceException
      */
     public void storeBlock(int id) throws BlockSourceException {
-    	storeBlock(id, (byte)-1);
+    	storeBlock(id, -1);
     }
-    public void storeBlock(int id, byte data) throws BlockSourceException {
+    public void storeBlock(int id, int data) throws BlockSourceException {
         if (!canStore) {
             throw new OutOfSpaceException(id);
         }
